// src/app/models/acte-mariage.model.ts
export class Acte_Mariage {
    id: number;
    num_acte: number;
    prenom_officier: string;
    nom_officier: string;
    prenom_marie: string;
    nom_marie: string;
    adress_marie: string;
    prenom_pere_du_marie: string;
    prenom_maman_marie: string;
    nom_maman_du_marie: string;
    prenom_fem: string;
    nom_fem: string;
    adress_fem: string;
    prenom_pere_fem: string;
    prenom_maman_fem: string;
    nom_maman_fem: string;
    option: string;
    regime_matrimoniale: string;
}